feat: Added app.py
